Split plugins sidebar into multiple files
