+++ src/components/backstage/plugins/Sidebar/NpmDetailsList.js
@@ -1,0 +1,90 @@
+import React from 'react';
+import kebabCase from 'lodash/kebabCase';
+import { Title } from 'components';
+import ContentLoader from 'react-content-loader';
+
+const DetailsListItem = ({ label, value, ...props }) => {
+  return (
+    <li
+      className="border-b border-gray-200 py-2 flex place-content-between"
+      id={`npm-detail-${kebabCase(label)}`}
+      {...props}
+    >
+      <span>{label}</span> <strong>{value}</strong>
+    </li>
+  );
+};
+
+const NpmDetailsList = ({ npmData, npmDataLoadingState }) => {
+  if (npmDataLoadingState === 'error') return null;
+  let inner;
+
+  if (npmDataLoadingState === 'loaded') {
+    inner = (
+      <div>
+        <ul className="mb-3">
+          <DetailsListItem label="Version" value={npmData.latestVersion} />
+          <DetailsListItem label="Downloads in last month" value={npmData.lastMonthDownloads} />
+          <DetailsListItem
+            label="Last published"
+            value={npmData.latestVersionPublishedAgo}
+            title={npmData.latestVersionPublishedTime}
+          />
+          <DetailsListItem
+            label="First published"
+            value={npmData.firstPublishedAgo}
+            title={npmData.firstPublishedTime}
+          />
+          <DetailsListItem label="Number of versions" value={npmData.numberOfVersions} />
+          <DetailsListItem label="License" value={npmData.license} />
+        </ul>
+
+        <span
+          className="flex place-content-between text-gray-400"
+          title={npmData.roadieLastUpdated}
+        >
+          <span className="italic">Last synced with NPM:</span>
+          <span>{npmData.lastSyncedAgo}</span>
+        </span>
+      </div>
+    );
+  } else {
+    inner = (
+      <ContentLoader
+        speed={2}
+        width={400}
+        height={240}
+        viewBox="0 0 400 240"
+        backgroundColor="#f3f3f3"
+        foregroundColor="#ecebeb"
+      >
+        <rect x="0" y="0" rx="3" ry="3" width="150" height="24" />
+        <rect x="220" y="0" rx="3" ry="3" width="150" height="24" />
+        <rect x="0" y="40" rx="3" ry="3" width="150" height="24" />
+        <rect x="220" y="40" rx="3" ry="3" width="150" height="24" />
+        <rect x="0" y="80" rx="3" ry="3" width="150" height="24" />
+        <rect x="220" y="80" rx="3" ry="3" width="150" height="24" />
+        <rect x="0" y="120" rx="3" ry="3" width="150" height="24" />
+        <rect x="220" y="120" rx="3" ry="3" width="150" height="24" />
+        <rect x="0" y="160" rx="3" ry="3" width="150" height="24" />
+        <rect x="220" y="160" rx="3" ry="3" width="150" height="24" />
+        <rect x="0" y="200" rx="3" ry="3" width="150" height="24" />
+        <rect x="220" y="200" rx="3" ry="3" width="150" height="24" />
+        <rect x="0" y="240" rx="3" ry="3" width="150" height="24" />
+        <rect x="220" y="240" rx="3" ry="3" width="150" height="24" />
+      </ContentLoader>
+    );
+  }
+
+  return (
+    <>
+      <div className="mb-4">
+        <Title>Plugin details</Title>
+      </div>
+
+      {inner}
+    </>
+  );
+};
+
+export default NpmDetailsList;
