+++ src/components/backstage/plugins/Sidebar/index.js
@@ -1,0 +1,1 @@
+export { default as Sidebar } from './Sidebar';
